Initial import of project files.
The first commit contains two filters: one which reports the number of HTTP
requests, and one which reports the latency of HTTP requests. It reports these
to a standard Prometheus CollectorRegistry.

There is also a MetricsService, which takes a CollectorRegistry and exposes
this over HTTP in the text format that Prometheus expects when scraping.

It also contains a TestServer, which shows all of this working together. This
is a useful testing ground, but won't remain in the library for long - I will
probably move it in to an examples directory.